implemented fetch posts, update post, delete post and like/dislike post controllers and all routes

+++ src/use-cases/delete-post/delete-post.ts
@@ -1,20 +1,30 @@
+import { TokenPayload, USER_ROLES } from '../../@types/types'
 import { PostsRepository } from '../../repositories/posts-repository'
 import { ResourceNotFoundError } from '../@errors/resource-not-found-error'
+import { UserNotAllowed } from '../@errors/user-not-alowed-error'
 
 interface DeletePostUseCaseRequest {
   id: string
+  user: TokenPayload
 }
 
 export class DeletePostUseCase {
   constructor(private postsRepository: PostsRepository) {}
 
-  async execute({ id }: DeletePostUseCaseRequest): Promise<void> {
+  async execute({ id, user }: DeletePostUseCaseRequest): Promise<void> {
     const post = await this.postsRepository.findById(id)
 
     if (!post) {
       throw new ResourceNotFoundError('Invalid post id')
     }
+    const isUserAdmin = user.role === USER_ROLES.ADMIN
+    const isUserTheCreator = user.id === post.creator_id
 
-    await this.postsRepository.delete(id)
+    if (isUserAdmin || isUserTheCreator) {
+      await this.postsRepository.delete(id)
+      return
+    }
+
+    throw new UserNotAllowed()
   }
 }
